rename files to Camel Case
