+++ src/components/introPageWrapper/styles.js
