+++ src/layouts/mapArea/index.js
@@ -1,9 +1,9 @@
 import React, { useRef, useEffect } from 'react';
 import mapboxgl from 'mapbox-gl';
 import 'mapbox-gl/dist/mapbox-gl.css';
-import MainPage from "../../components/main-page-wrapper";
+import MainPage from "../../components/mainPageWrapper";
 import Map from "./map";
-import OrderArea from "./order/order-area"
+import OrderArea from "./order/orderArea"
 
 
 const styles = {
