+++ src/layouts/mapArea/order/orderArea.js
@@ -3,9 +3,9 @@ import { connect } from 'react-redux';
 import { Container, Grid, Button, Typography, Paper } from '@material-ui/core';
 import withStyles from '@material-ui/core/styles/withStyles';
 import { getIsPaymentMethodReceived } from '../../../modules/card';
-import OrderForm from './order-form';
-import NeedCard from "./need-card-msg";
-import OrderSuccess from "./order-success";
+import OrderForm from './orderForm';
+import NeedCard from "./needCardMsg";
+import OrderSuccess from "./orderSuccess";
 import styles from './styles';
 
 export const OrderArea = ({ classes, isPaymentMethodReceived, flagOrderSuccessMsgInit=false}) => {
