+++ src/layouts/mapArea/order/orderArea.test.js
@@ -6,13 +6,13 @@ import {act, render, fireEvent, waitFor, wait} from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
 import {rootReducer} from '../../../modules'
-import OrderArea from "./order-area";
+import OrderArea from "./orderArea";
 import { createStore } from 'redux'
 import {Provider } from 'react-redux';
 // import OrderForm from "./OrderForm";
-import NeedCard from "./need-card-msg";
-import OrderSuccess from "./order-success";
-import {renderWithProviders} from '../../../utils/test-utils'
+import NeedCard from "./needCardMsg";
+import OrderSuccess from "./orderSuccess";
+import {renderWithProviders} from '../../../utils/testUtils'
 
 
 const store = createStore(rootReducer, {})
