+++ src/layouts/profile/paymentForm.test.js
@@ -8,14 +8,14 @@ import {rootReducer} from '../../modules/'
 
 import {createStore} from 'redux'
 
-import {PaymentForm} from "./payment-form";
-import {renderWithProviders, renderWithTheme} from '../../utils/test-utils'
+import {PaymentForm} from "./paymentForm";
+import {renderWithProviders, renderWithTheme} from '../../utils/testUtils'
 
 
 const store = createStore(rootReducer, {})
 describe("order form ", () => {
-    const confirmCardSaved=jest.fn();
-    const saveCard=jest.fn();
+    const confirmCardSaved = jest.fn();
+    const saveCard = jest.fn();
 
     beforeEach(() => {
         jest.resetAllMocks();
@@ -24,7 +24,7 @@ describe("order form ", () => {
     it("renders correctly", () => {
         const {getByTestId, queryByTestId, getByText} = renderWithTheme(
             <PaymentForm classes={{}} confirmCardSaved={jest.fn()}/>
-            );
+        );
         // expect(getByTestId("order-form")).toBeTruthy();
         expect(getByText("Номер карты")).toBeTruthy();
         expect(getByText("Срок действия")).toBeTruthy();
@@ -40,15 +40,19 @@ describe("order form ", () => {
                 saveCard={confirmCardSaved}/>
         );
 
-        wait(() => {fireEvent.submit(
-            getByTestId('save-card-form'),
-            {target: {
-                    card_number: '0000 0000 0000 0000',
-                    card_date: '12/12',
-                    card_name: 'vasia',
-                    card_cvc: '123'
-                }}
-            )});
+        wait(() => {
+            fireEvent.submit(
+                getByTestId('save-card-form'),
+                {
+                    target: {
+                        card_number: '0000 0000 0000 0000',
+                        card_date: '12/12',
+                        card_name: 'vasia',
+                        card_cvc: '123'
+                    }
+                }
+            )
+        });
         wait(() => {
             expect(confirmCardSaved).toHaveBeenCalled()
         });
@@ -59,7 +63,7 @@ describe("order form ", () => {
 
     it("card have not been saved", async () => {
 
-        const {getByTestId, queryByTestId, getByText} = renderWithTheme(
+        const {getByTestId} = renderWithTheme(
             <PaymentForm
                 classes={{}}
                 confirmCardSaved={confirmCardSaved}
